Document our "About" and "Server rules"
The Mastodon software provides a public interface for an instance's
description and summary of their rules. It is customary to use this
space to demonstrate our commitment to moderate our instance, both to
our users and to those who interact with them.

This commit introduces these rules as Markdown files into the Mastodon
repo. This is not the final step -- the agreed-upon text must then be
pasted into a textarea. This is simply a place for us to discuss and
archive the text, in a format and UX that is familiar.